requirements.txt
Django==5.2
requests==2.31.0
gunicorn==21.2.0
python-dotenv